maint: rename dm-handler → dispatcher-dm-handler
The file implements the dispatcher's DM control surface (watch/unwatch/watch-pr
for APM ↔ dispatcher flow) — the old name `dm-handler` looked like a generic
handler and nearly caused an accidental deletion. Rename makes the scope clear.

Renames both src and test files; updates all import paths, comments, and log
prefixes throughout. No logic changes.

## Human-Claude Interaction Log

### Human prompts (VERBATIM - include typos, informal language, COMPLETE text):
1. "/worker roost 395 AvesAlight/roost maint/395-rename-dm-handler alex"
   → Claude: joined #roost-issue-395, read issue, posted plan, did full grep sweep, then executed the rename after lead-pm approval
